+++ test/models/disco_app/has_metafields_test.rb
@@ -15,7 +15,7 @@ class DiscoApp::HasMetafieldsTest < ActiveSupport::TestCase
 
   test 'can write metafields with a single namespace' do
     stub_api_request(:put, "#{@shop.admin_url}/products/#{@product.id}.json", 'widget_store/products/write_metafields_single_namespace')
-    assert @shop.temp { @product.write_metafields(
+    assert @shop.with_api_context { @product.write_metafields(
       namespace1: {
         key1: 'value1',
         key2: 2
@@ -25,7 +25,7 @@ class DiscoApp::HasMetafieldsTest < ActiveSupport::TestCase
 
   test 'can write metafields with multiple namespaces' do
     stub_api_request(:put, "#{@shop.admin_url}/products/#{@product.id}.json", 'widget_store/products/write_metafields_multiple_namespaces')
-    assert @shop.temp { @product.write_metafields(
+    assert @shop.with_api_context { @product.write_metafields(
       namespace1: {
         n1key1: 'value1',
         n1key2: 2